chore: update pre-commit configuration and enhance type safety in environment files
- Added a new `pyright` configuration for Python files in the `POMDPPlanners` directory to improve type checking.
- Updated type annotations and added type ignores in various environment files to enhance type safety and clarity, particularly in plotting and object creation.
- Refactored the `HyperParameterOptimizer` simulation examples to utilize temporary directories for caching, improving test isolation and cleanup.

These changes enhance the maintainability and robustness of the POMDPPlanners library.

+++ POMDPPlanners/core/cost.py
@@ -35,7 +35,7 @@ def belief_expectation_cost_particle_belief(
     costs = np.array(
         [-env.reward(belief.particles[i], action) for i in range(len(belief.particles))]
     )
-    cost_: float = np.sum(costs * belief.normalized_weights)
+    cost_: float = float(np.sum(costs * belief.normalized_weights))
 
     return cost_
